+++ src/services/apiClient.ts
@@ -1,4 +1,4 @@
-import type { CodeData, DiffData, FileData, FileListItem } from '../types';
+import type { CodeData, DiffData, FileData, FileListItem } from '../../shared/types';
 const API_BASE_URL = import.meta.env.VITE_API_BASE_URL || 'http://localhost:3000/api';
 
 class ApiClient {
